feat: virtualize plugins and app-config

+++ packages/core/src/node/vixt.ts
@@ -2,13 +2,15 @@ import type { VixtConfigLayer, VixtOptions } from './config'
 import type { VixtModule } from './module'
 import type { LoadConfigOptions } from 'c12'
 
-import { cac } from 'cac'
+import { env } from 'node:process'
+
 import defu from 'defu'
 import fs from 'fs-extra'
 
-import { applyLayers, config, typescript } from '.'
-import { loadVixtConfig } from './config'
-import { applyLayerModules, defineVitePlugin } from './module'
+import { applyLayers, loadVixtConfig } from './config'
+import { loadCLIOptions } from './env'
+import { applyLayerModules, defineVitePlugin, installModule } from './module'
+import { builtinModules } from './modules'
 
 export interface Vixt {
   options: VixtOptions
@@ -17,18 +19,18 @@ export interface Vixt {
 }
 
 export async function loadVixt(opts?: LoadConfigOptions<VixtOptions>) {
-  const result = await loadVixtConfig(defu(opts, {
-    defaults: {
-      modules: [config, typescript],
-    },
-  }))
+  const result = await loadVixtConfig(opts)
+
+  const cliOptions = loadCLIOptions()
 
   // remove buildDir
-  const parsedArgv = cac().parse()
-  const isForce = !!parsedArgv.options.force
-  if (isForce) {
-    fs.removeSync(result.config.buildDir!)
-  }
+  cliOptions.force && fs.removeSync(result.config.buildDir!)
+
+  // assign vixt.options.debug
+  result.config.debug = !!cliOptions.debug
+
+  // assign vixt.options.dev
+  result.config.dev = env.NODE_ENV !== 'production'
 
   // apply layers
   result.layers = applyLayers(result.layers ?? [], result.config)
@@ -39,7 +41,7 @@ export async function loadVixt(opts?: LoadConfigOptions<VixtOptions>) {
   const vixt: Vixt = {
     options: result.config,
     _layers: result.layers ?? [],
-    _modules: [...result.config.modules ?? [], ...layerModules],
+    _modules: [...builtinModules, ...result.config.modules ?? [], ...layerModules],
   }
 
   return vixt
@@ -49,7 +51,7 @@ export function createVixtPlugin(loadOptions: LoadConfigOptions<VixtOptions>) {
   return defineVitePlugin<VixtOptions>(async (vixtOptions) => {
     const vixt = await loadVixt(defu({ defaults: vixtOptions }, loadOptions))
 
-    const plugins = vixt._modules.map(module => module({}, vixt))
+    const plugins = vixt._modules.map(module => installModule(module, {}, vixt))
 
     return plugins
   })
